Create competitions context
