+++ test/actions/paper_vote_actions_test.exs
@@ -1,7 +1,9 @@
 defmodule ApiWeb.PaperVoteActionsTest do
   use Api.DataCase
 
-  alias ApiWeb.{CompetitionActions, PaperVoteActions, TeamActions, Team}
+  alias Api.Competitions
+  alias Api.Competitions.Team
+  alias ApiWeb.PaperVoteActions
 
   setup do
     member = create_user()
@@ -25,8 +27,8 @@ defmodule ApiWeb.PaperVoteActionsTest do
   end
 
   test "create after end", %{category: c, admin: a} do
-    CompetitionActions.start_voting()
-    CompetitionActions.end_voting()
+    Competitions.start_voting()
+    Competitions.end_voting()
 
     :already_ended = PaperVoteActions.create(c, a)
   end
@@ -34,7 +36,7 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "redeem", %{category: c, admin: a, member: m, team: t} do
     {:ok, p} = PaperVoteActions.create(c, a)
     [t] = make_teams_eligible([t])
-    CompetitionActions.start_voting()
+    Competitions.start_voting()
 
     {:ok, _} = PaperVoteActions.redeem(p, t, m, a)
   end
@@ -42,7 +44,7 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "redeem not twice", %{category: c, admin: a, member: m, team: t} do
     {:ok, p} = PaperVoteActions.create(c, a)
     [t] = make_teams_eligible([t])
-    CompetitionActions.start_voting()
+    Competitions.start_voting()
 
     {:ok, p} = PaperVoteActions.redeem(p, t, m, a)
     :already_redeemed = PaperVoteActions.redeem(p, t, m, a)
@@ -51,7 +53,7 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "redeem not annulled", %{category: c, admin: a, member: m, team: t} do
     {:ok, p} = PaperVoteActions.create(c, a)
     [t] = make_teams_eligible([t])
-    CompetitionActions.start_voting()
+    Competitions.start_voting()
 
     {:ok, p} = PaperVoteActions.annul(p, a)
     :annulled = PaperVoteActions.redeem(p, t, m, a)
@@ -60,7 +62,7 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "redeem not eligible", %{category: c, admin: a, member: m, team: t} do
     {:ok, p} = PaperVoteActions.create(c, a)
     # Notice I'm not making teams eligible
-    CompetitionActions.start_voting()
+    Competitions.start_voting()
 
     :team_not_eligible = PaperVoteActions.redeem(p, t, m, a)
   end
@@ -68,9 +70,9 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "redeem disqualified", %{category: c, admin: a, member: m, team: t} do
     {:ok, p} = PaperVoteActions.create(c, a)
     [t] = make_teams_eligible([t])
-    CompetitionActions.start_voting()
+    Competitions.start_voting()
 
-    TeamActions.disqualify(t.id, a)
+    Competitions.disqualify_team(t.id, a)
     t = Repo.get!(Team, t.id)
 
     :team_disqualified = PaperVoteActions.redeem(p, t, m, a)
@@ -86,8 +88,8 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "redeem after end", %{category: c, admin: a, member: m, team: t} do
     {:ok, p} = PaperVoteActions.create(c, a)
     [t] = make_teams_eligible([t])
-    CompetitionActions.start_voting()
-    CompetitionActions.end_voting()
+    Competitions.start_voting()
+    Competitions.end_voting()
 
     :already_ended = PaperVoteActions.redeem(p, t, m, a)
   end
@@ -100,8 +102,8 @@ defmodule ApiWeb.PaperVoteActionsTest do
   test "annul after end", %{category: c, admin: a} do
     {:ok, p} = PaperVoteActions.create(c, a)
 
-    CompetitionActions.start_voting()
-    CompetitionActions.end_voting()
+    Competitions.start_voting()
+    Competitions.end_voting()
 
     :already_ended = PaperVoteActions.annul(p, a)
   end
