+++ test/actions/user_actions_test.exs
@@ -2,13 +2,13 @@ defmodule ApiWeb.UserActionTest do
   use Api.DataCase
 
   alias Api.Accounts
-  alias ApiWeb.CompetitionActions
+  alias Api.Competitions
 
 
   test "able_to_vote checked in" do
     u = create_user()
 
-    CompetitionActions.start_voting()
+    Competitions.start_voting()
 
     :already_started = Accounts.toggle_checkin(u.id, true)
   end
