+++ test/models/membership_test.exs
@@ -1,23 +1,23 @@
-defmodule ApiWeb.TeamMemberTest do
+defmodule ApiWeb.MembershipTest do
   use Api.DataCase
 
-  alias ApiWeb.{TeamMember}
+  alias Api.Competitions.Membership
 
   @valid_attrs %{user_id: Ecto.UUID.generate(), team_id: Ecto.UUID.generate()}
   @invalid_attrs %{}
 
   test "changeset with valid attributes" do
-    changeset = TeamMember.changeset(%TeamMember{}, @valid_attrs)
+    changeset = Membership.changeset(%Membership{}, @valid_attrs)
     assert changeset.valid?
   end
 
   test "changeset with invalid attributes" do
-    changeset = TeamMember.changeset(%TeamMember{}, @invalid_attrs)
+    changeset = Membership.changeset(%Membership{}, @invalid_attrs)
     refute changeset.valid?
   end
 
   test "changeset with no attributes" do
-    changeset = TeamMember.changeset(%TeamMember{})
+    changeset = Membership.changeset(%Membership{})
     refute changeset.valid?
   end
 end
